backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

